All user Authentications happen in getPollAndAuthParticipant
Previously only "AuthLink" Authentications happened during getPollAndAuthParticipant, while authentications from the jwt where handled in app.tsx by calling a redux action that sends an http request to the REST API.
Now, the jwt is still set in app.tsx as the http reuqest header and requests to etPollAndAuthParticipant without the AuthLink and with an http request header will respond with an empty string as token(""). 
If the token is such an empty string it is not set new in the setAuthTokenAndUser action.